+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- `axi_to_detailed_mem`: Ensure proper propagation or `err` and `exokay` signals.
+
 ## 0.39.0 - 2023-07-20
 
 ### Added